Switch to loading bootstrap CSS and JS through Webpack
This makes the build even worse, requiring first a babel step and then
the webpack step! But that's just temporary, next goal is to run the
Babel step through Webpack as well, which leaves us at roughly the same
awfulness as before. I couldn't quite get bootstrap-icons to work, so
kicking that down the road as well